+++ src/lib/pages/instantiate/instantiate.tsx
@@ -44,6 +44,7 @@ import {
 import { FooterCTA } from "lib/components/layouts";
 import { CodeSelectSection } from "lib/components/select-code";
 import { Stepper } from "lib/components/stepper";
+import { UserDocsLink } from "lib/components/UserDocsLink";
 import WasmPageContainer from "lib/components/WasmPageContainer";
 import { useSchemaStore } from "lib/providers/store";
 import { useTxBroadcast } from "lib/providers/tx-broadcast";
@@ -392,9 +393,16 @@ const Instantiate = ({ onComplete }: InstantiatePageProps) => {
           DEPLOY NEW CONTRACT
         </Text>
         <Stepper mode="deploy" currentStep={2} />
-        <Heading as="h5" variant="h5" my={12}>
-          Instantiate new contract
-        </Heading>
+        <Flex direction="column" alignItems="center" my={12}>
+          <Heading as="h5" variant="h5">
+            Instantiate new contract
+          </Heading>
+          <UserDocsLink
+            mt={2}
+            cta="View Instantiate Guideline"
+            href="cosmwasm/deploy-contract#instantiate-contract-from-code"
+          />
+        </Flex>
         <ConnectWalletAlert
           subtitle="You need to connect your wallet to perform this action"
           mb={6}