+++ src/lib/pages/proposal-details/components/VpPercentThreshold.tsx
@@ -4,7 +4,7 @@ import { Fragment } from "react";
 import { normalizeVotesInfo } from "../utils";
 import { CustomIcon } from "lib/components/icon";
 import type { ProposalVotesInfo } from "lib/types";
-import { d0Formatter, divWithDefault, formatPrettyPercent } from "lib/utils";
+import { d0Formatter, formatPrettyPercent } from "lib/utils";
 
 import { VpPercentCard } from "./VpPercentCard";
 
@@ -17,25 +17,25 @@ export const VpPercentThreshold = ({
   votesInfo,
   isCompact,
 }: VpPercentThresholdProps) => {
-  const { yes, no, noWithVeto, nonAbstainVotes } =
+  const { yesNonRatio, noNonRatio, noWithVetoNonRatio } =
     normalizeVotesInfo(votesInfo);
 
   const options = [
     {
       option: "Yes",
-      ratio: divWithDefault(yes, nonAbstainVotes, 0),
+      ratio: yesNonRatio,
       votingPower: votesInfo.yes,
       color: "success.main",
     },
     {
       option: "No",
-      ratio: divWithDefault(no, nonAbstainVotes, 0),
+      ratio: noNonRatio,
       votingPower: votesInfo.no,
       color: "error.main",
     },
     {
       option: "No with veto",
-      ratio: divWithDefault(noWithVeto, nonAbstainVotes, 0),
+      ratio: noWithVetoNonRatio,
       votingPower: votesInfo.noWithVeto,
       color: "error.dark",
     },
@@ -67,7 +67,7 @@ export const VpPercentThreshold = ({
           </Flex>
           <Flex direction="column" align="end">
             <Text variant="body2" color="text.main">
-              {formatPrettyPercent(ratio.toNumber(), 2, true)}
+              {formatPrettyPercent(ratio, 2, true)}
             </Text>
             <Text variant="body3" color="text.dark">
               ({d0Formatter(votingPower, "0")})