+++ src/lib/pages/proposal-details/components/proposal-overview/proposal-period-overview/VotingOverview.tsx
@@ -49,11 +49,13 @@ const VotingOverviewBody = ({
 
   return (
     <>
-      <VotingOverviewQuorum
-        proposalData={proposalData}
-        params={params}
-        votesInfo={votesInfo}
-      />
+      {proposalData.status === ProposalStatus.VOTING_PERIOD && (
+        <VotingOverviewQuorum
+          proposalData={proposalData}
+          params={params}
+          votesInfo={votesInfo}
+        />
+      )}
       <VotingOverviewThreshold
         proposalData={proposalData}
         params={params}
@@ -75,7 +77,7 @@ const VotingOverviewBody = ({
           })
         }
       >
-        View Full Vote Details
+        View Full Voting Details
       </Button>
     </>
   );