+++ src/lib/pages/proposal-details/components/vote-details/voting-period/votes-table/ProposalVotesTableRow.tsx
@@ -1,10 +1,9 @@
 import { Button, Flex, Grid, Text } from "@chakra-ui/react";
 
-import { Answer } from "../table/Answer";
 import { Voter } from "../table/Voter";
 import { useMobile } from "lib/app-provider";
 import { ExplorerLink } from "lib/components/ExplorerLink";
-import { TableRow } from "lib/components/table";
+import { Answer, TableRow } from "lib/components/table";
 import { useOpenTxTab } from "lib/hooks";
 import type { ProposalVote } from "lib/types";
 import { dateFromNow, formatUTC } from "lib/utils";
@@ -65,7 +64,13 @@ export const ProposalVotesTableRow = ({
         <Voter proposalVote={proposalVote} />
       </TableRow>
       <TableRow>
-        <Answer proposalVote={proposalVote} />
+        <Answer
+          isVoteWeighted={proposalVote.isVoteWeighted}
+          yes={proposalVote.yes}
+          no={proposalVote.no}
+          noWithVeto={proposalVote.noWithVeto}
+          abstain={proposalVote.abstain}
+        />
       </TableRow>
       {fullVersion && !isMobile && (
         <>