+++ src/lib/pages/validator-details/components/bonded-token-changes/VotingPowerChart.tsx
@@ -11,6 +11,7 @@ import { useValidatorHistoricalPowers } from "lib/services/validatorService";
 import type { AssetInfos, Option, Token, U, ValidatorAddr } from "lib/types";
 import {
   formatHHmm,
+  formatUTC,
   formatUTokenWithPrecision,
   getTokenLabel,
 } from "lib/utils";
@@ -43,39 +44,13 @@ export const VotingPowerChart = ({
     ? assetInfos?.[singleStakingDenom]
     : undefined;
 
-  if (isMobileOverview)
-    return (
-      <Flex
-        backgroundColor="gray.900"
-        p={4}
-        rounded={8}
-        w="100%"
-        justifyContent="space-between"
-        alignItems="center"
-        onClick={onViewMore}
-      >
-        <VotingPowerChartDetails
-          historicalPowers={historicalPowers}
-          singleStakingDenom={singleStakingDenom}
-          assetInfo={assetInfo}
-        />
-        <CustomIcon boxSize={6} m={0} name="chevron-right" color="gray.600" />
-      </Flex>
-    );
+  const currency = singleStakingDenom
+    ? `${getTokenLabel(singleStakingDenom, assetInfo?.symbol)}`
+    : "";
 
   const labels = historicalPowers.items.map((item) =>
     formatHHmm(item.hourRoundedTimestamp as Date)
   );
-
-  const dateLabels = labels.map((label) => {
-    const [hours, minutes] = label.split(":");
-    const date = new Date();
-    date.setHours(parseInt(hours, 10));
-    date.setMinutes(parseInt(minutes, 10));
-    date.setSeconds(0);
-
-    return date.toString().replace(" GMT", "<br>GMT");
-  });
 
   const dataset = {
     data: historicalPowers.items.map((item) => item.votingPower.toNumber()),
@@ -94,10 +69,6 @@ export const VotingPowerChart = ({
     pointHoverBorderColor: "#D8BEFC",
   };
 
-  const currency = singleStakingDenom
-    ? `${getTokenLabel(singleStakingDenom, assetInfo?.symbol)}`
-    : "";
-
   const customizeTooltip = (tooltip: TooltipModel<"line">) => {
     const { raw, dataIndex } = tooltip.dataPoints[0];
 
@@ -106,6 +77,10 @@ export const VotingPowerChart = ({
       assetInfo?.precision ?? 0,
       false,
       2
+    );
+
+    const formattedDate = formatUTC(
+      historicalPowers.items[dataIndex].hourRoundedTimestamp
     );
 
     return `
@@ -117,12 +92,29 @@ export const VotingPowerChart = ({
           <p style="font-size: 16px; color: #F7F2FE; white-space: nowrap;">${formattedAmount} ${currency}</p>
         </div>
         <hr style="margin-top: 8px; color: #68688A;"/>
-        <p style="margin-top: 8px; font-size: 12px; color: #F7F2FE; white-space: nowrap;">${dateLabels[dataIndex]}</p>
+        <p style="margin-top: 8px; font-size: 12px; color: #F7F2FE; white-space: nowrap;">${formattedDate}</p>
       </div>
     `;
   };
 
-  return (
+  return isMobileOverview ? (
+    <Flex
+      backgroundColor="gray.900"
+      p={4}
+      rounded={8}
+      w="100%"
+      justifyContent="space-between"
+      alignItems="center"
+      onClick={onViewMore}
+    >
+      <VotingPowerChartDetails
+        historicalPowers={historicalPowers}
+        singleStakingDenom={singleStakingDenom}
+        assetInfo={assetInfo}
+      />
+      <CustomIcon boxSize={6} m={0} name="chevron-right" color="gray.600" />
+    </Flex>
+  ) : (
     <Flex
       direction={{
         lg: "row",