+++ src/lib/pages/validator-details/components/validator-top/ValidatorStats.tsx
@@ -5,8 +5,8 @@ import {
   useValidatorDelegators,
   useValidatorStakingProvisions,
 } from "lib/services/validatorService";
-import type { Option, ValidatorAddr } from "lib/types";
-import { formatPrettyPercent } from "lib/utils";
+import type { Option, Ratio, ValidatorAddr } from "lib/types";
+import { divWithDefault, formatPrettyPercent } from "lib/utils";
 
 const StatWithLabel = ({
   label,
@@ -37,7 +37,7 @@ const StatWithLabel = ({
 
 interface ValidatorStatsProps {
   validatorAddress: ValidatorAddr;
-  commissionRate: number;
+  commissionRate: Ratio<number>;
   totalVotingPower: Big;
   singleStakingDenom: Option<string>;
 }
@@ -49,15 +49,17 @@ export const ValidatorStats = ({
   singleStakingDenom,
 }: ValidatorStatsProps) => {
   const { data: stakingProvisions, isLoading: isStakingProvisionsLoading } =
-    useValidatorStakingProvisions();
+    useValidatorStakingProvisions(!!singleStakingDenom);
   const { data: delegations, isLoading: isDelegationsLoading } =
     useValidatorDelegators(validatorAddress);
 
   const estimatedApr = stakingProvisions
     ? formatPrettyPercent(
-        stakingProvisions.stakingProvisions
-          .div(totalVotingPower.mul(1 - commissionRate))
-          .toNumber(),
+        divWithDefault(
+          stakingProvisions.stakingProvisions,
+          totalVotingPower.mul(1 - commissionRate),
+          0
+        ).toNumber() as Ratio<number>,
         2,
         true
       )