+++ src/lib/types/proposal.ts
@@ -6,6 +6,7 @@ import type {
   BechAddr,
   Nullable,
   Option,
+  Ratio,
   TokenWithValue,
   Validator,
 } from "lib/types";
@@ -84,14 +85,14 @@ export interface ProposalParams<
   minInitialDepositRatio: number;
   maxDepositPeriod: string;
   votingPeriod: string;
-  vetoThreshold: number;
-  quorum: number;
-  threshold: number;
+  vetoThreshold: Ratio<number>;
+  quorum: Ratio<number>;
+  threshold: Ratio<number>;
   // expedited
   expeditedVotingPeriod?: string;
-  expeditedThreshold?: number;
+  expeditedThreshold?: Ratio<number>;
   expeditedMinDeposit?: T[];
-  expeditedQuorum?: number; // only in sei
+  expeditedQuorum?: Ratio<number>; // only in sei
   // emergency - only in initia
   emergencyMinDeposit?: T[];
   emergencyTallyInterval?: string;
@@ -114,6 +115,7 @@ interface Message {
 
 export interface ProposalData<T extends Coin | TokenWithValue = TokenWithValue>
   extends Proposal {
+  failedReason: string;
   createdHeight: Nullable<number>;
   createdTimestamp: Nullable<Date>;
   createdTxHash: Nullable<string>;
@@ -153,12 +155,12 @@ export interface ProposalValidatorVote extends ProposalVote {
   rank: number;
 }
 
-export enum ProposalValidatorVoteType {
+export enum ProposalVoteType {
   ALL = "all",
   YES = "yes",
   NO = "no",
   NO_WITH_VETO = "no_with_veto",
   ABSTAIN = "abstain",
   WEIGHTED = "weighted",
-  DID_NOT_VOTE = "did_not_vote",
+  DID_NOT_VOTE = "did_not_vote", // only for validator votes
 }